+++ scripts/tests/docs-link/cases/06_unlink_strips_parent_and_http.sh
@@ -31,7 +31,7 @@ git -C "$SYSTEM" init -q
 git -C "$COMPANY" remote add origin "https://github.com/example/company-ea.git"
 git -C "$SYSTEM" remote add origin "https://example.com/org/sys-foo.git"
 
-cp -R "$ROOT_DIR/company/system-SYSNAME" "$COMPANY/docs/system-SYSNAME"
+cp -R "$ROOT_DIR/company/system-slots" "$COMPANY/docs/system-slots"
 printf '%s\n' 'links: []' >"$SYSTEM/docs/knowledge-links.yaml"
 
 cat >"$COMPANY/.docsconfig" <<EOF